requirements.txt
astroid==2.12.13
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
customtkinter==5.1.2
darkdetect==0.8.0
decorator==5.1.1
dill==0.3.6
distlib==0.3.6
filelock==3.9.0
flake8==6.0.0
flake8-docstrings==1.6.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
mock==4.0.3
mypy==0.991
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
pbr==5.11.1
pep8-naming==0.13.2
Pillow==9.4.0
platformdirs==2.5.4
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pydocstyle==6.1.1
pyflakes==3.0.1
pyftdi==0.54.0
pylint==2.15.7
pyparsing==3.0.9
pyserial==3.5
pytest==7.2.0
pytest-dotenv==0.5.2
pytest-mock==3.10.0
python-dotenv==0.21.0
python-mpv==1.0.1
pytube==12.1.0
pyusb==1.2.1
requests==2.28.1
retry==0.9.2
setproctitle==1.3.2
snowballstemmer==2.2.0
stevedore==4.1.1
tomlkit==0.11.6
types-retry==0.9.9
typing_extensions==4.4.0
urllib3==1.26.13
virtualenv==20.17.1
virtualenv-clone==0.5.7
wrapt==1.14.1
